requirements.txt
dbus_python==1.2.18
pytest==7.2.1
requests==2.25.1
